Add current features hook and extend selected feature hook
