+++ react/src/components/AssetDetail/AssetDetail.test.tsx
@@ -2,11 +2,7 @@ import React, { act } from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import AssetDetail from './AssetDetail';
 import { featureCollection } from '@hazmapper/__fixtures__/featuresFixture';
-import {
-  mockImgFeature,
-  mockVideoFeature,
-  mockPointFeature,
-} from '@hazmapper/__fixtures__/assetDetailFixture';
+import { mockImgFeature } from '@hazmapper/__fixtures__/featuresFixture';
 import { useAppConfiguration, useFeatureSelection } from '@hazmapper/hooks';
 
 jest.mock('@hazmapper/hooks', () => ({
